add clear products
