+++ src/components/card/CardModal.vue
@@ -3,12 +3,7 @@
   .card-header
     h2.card-title Корзина
     button.card_button--close(@click="closeCard") X 
-  .card-products 
-    p.card-products-count(v-if="productsInCard.length === 0")
-    p.card-products-count(v-else-if="productsInCard.length < 5 ") {{ productsInCard.length}} {{ productsInCard.length === 1 ? 'товар' : 'товара' }}
-    p.card-products-count(v-else) {{ productsInCard.length}} товаров
-    button.card-products-clear(@click="clearCard")  очистить список
-    <ProductCardList />
+  <ProductCardList />
   .product-total
     p.product-total-title Итого
     p.product-total-price 0₽
@@ -35,15 +30,10 @@ export default defineComponent({
       store.commit("setIsCardOpen", false);
     };
 
-    const clearCard = () => {
-      store.commit("setProductsInCard", []);
-    };
-
     return {
       closeCard,
       isCardOpen,
       productsInCard,
-      clearCard,
     };
   },
 });
@@ -76,6 +66,10 @@ export default defineComponent({
   }
 }
 
+.card-products {
+  flex: 1;
+}
+
 .card-title {
   font-weight: 500;
   font-size: $fz-30px;
@@ -85,5 +79,40 @@ export default defineComponent({
 
 .card-header {
   @include flex-between;
+}
+
+.product-total {
+  padding-block: 80px;
+  @include flex-between;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-template-rows: auto auto;
+
+  &-price {
+    grid-column: 1 / 2;
+    grid-row: 2 / 3;
+    font-weight: 500;
+    font-size: $fz-30px;
+    letter-spacing: -0.02em;
+  }
+
+  &-title {
+    grid-column: 1 / 2;
+    grid-row: 1 / 2;
+    text-transform: capitalize;
+  }
+
+  &-button {
+    padding-block: 20px;
+    padding-inline: 58px;
+    grid-column: 2 / 3;
+    grid-row: 1 / -1;
+    background-color: $color-brand;
+    border-radius: 4px;
+    font-weight: 500;
+    font-size: $fz-12px;
+    letter-spacing: 0.06em;
+    text-transform: uppercase;
+  }
 }
 </style>
